feat: conntrack delete message type and attributes
Implemented the `Delete`/`IPCTNL_MSG_CT_DELETE` conntrack message type and the following attributes, which are required to describe a conntrack entry for deletion:

* mark
* status
* timeout
* tuple_reply

Added tests for deleting TCP/IPv4 and UDP/IPv6 conntrack entries.

Added ProtoFamily, an enum that represents a protocol family in the Netfilter header.

The bitflags dependency was also bumped to v2.10.0.

Signed-off-by: Shivang K Raghuvanshi <shivangraghuvanshi2005@gmail.com>